dashboard/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
# sqlite3 is part of the Python standard library, so it does not need to be pip installed.
